+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 0.1.0 / 2014-03-10
 ------------------
 * added browser tests
+* added support for `coinstring()` method to input a type `Array` or `Uint8Array`
 
 0.0.1 / 2014-03-07
 ------------------